Finalized training code. Merged transformers and ru_transformers scripts.
