+++ train/yt_encoder.py
@@ -13,8 +13,6 @@ class YTEncoder(PreTrainedTokenizer):
     def_name = 'encoder.model'
     def __init__(self, filename, *inputs, **kwargs):
         super().__init__(*inputs, **kwargs)
-        self.max_len_single_sentence = 1024 # no default special tokens - you can update this value if you add special tokens
-        self.max_len_sentences_pair = 1024 # no default special tokens - you can update this value if you add special tokens
 
         if os.path.isdir(filename): filename = os.path.join(filename, self.def_name)
 
@@ -46,6 +44,9 @@ class YTEncoder(PreTrainedTokenizer):
     @classmethod
     def from_pretrained(cls, *inputs, **kwargs):
         return cls(*inputs, **kwargs)
+
+    def build_inputs_with_special_tokens(self, token_ids):
+        return token_ids
 
     def add_special_tokens_single_sentence(self, token_ids):
         return token_ids
